+++ Modules/Message/Database/Migrations/2021_07_23_210631_create_registrant_messages_table.php
@@ -15,7 +15,7 @@ class CreateRegistrantMessagesTable extends Migration
     {
         Schema::create('registrant_messages', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('registrant_id')->nullable();
+            $table->bigInteger('registrant_id')->nullable();
             $table->boolean('register_pass_message_sent')->nullable()->default(0);
             $table->boolean('requirements_pass_message_sent')->nullable()->default(0);
             $table->boolean('test_pass_message_sent')->nullable()->default(0);
