+++ Modules/Registrant/Database/Migrations/2021_08_09_172312_add_installment_id_to_registrants.php
@@ -14,7 +14,7 @@ class AddInstallmentIdToRegistrants extends Migration
     public function up()
     {
         Schema::table('registrants', function (Blueprint $table) {
-            $table->integer('installment_id')->nullable();
+            $table->bigInteger('installment_id')->nullable();
         });
     }
